Update worker instance connections during thorctl logs execution
Now that THOR's task queue has autoscaling, it's a bit more painful
get logs. A user might want to start fetching logs right after
submitting a job, but there are no instances to connect to to stream
logs from at that point.

So, when running thorctl logs, we need to check to see if new workers
have been added. We also need to be very tolerant of problems
connecting to new workers - when they first come online, they might
not have an external IP address, and then might not have an SSH server
running. We can just retry these cases, though.

In the course of implementing this, it was simpler to refactor the
existing sshconn.py code significantly.