Announce when tasks have completed to a PubSub topic
Add fields to JobManifests:
 - incomplete_tasks: a list of Task IDs of tasks that have not
   completed for the job.
 - pubsub_topic: An optional string name of a PubSub topic to publish
   to when all tasks for the job are done.
 - update_time: The most recent time the JobManifest was modified.

Add logic to thor.taskqueues.client.Worker to update the JobManifest
for a task's job when a task is completed. The completed task is
removed from JobManifest's 'incomplete_tasks' list.

This update risks races, so the function which updates a JobManifest
uses optimistic locking based on the 'generation' of the JobManifest's
blob in Cloud Storage.

When the incomplete_tasks has just one entry, and the Worker removes
it, the Worker knows it must have been the last task worker, so it can
announce the completion of the job. It publishes the JobManifest to
the pubsub topic specified.

Moved the thor.taskqueues.tasks.Task.mark_success and .mark_failed
methods into thor.taskqueues.client.Worker, since the Worker is a more
active component while the Task represents simple data.

Add logic to thor.taskqueues.client.Client to provide a PubSub topic
name when launching a job.

Add google-cloud-pubsub as a new dependency.

Add new arguments to submitTHORJob.py to allow setting the pubsub
topic for a job.

Update the integration test in thor/tests/test_taskqueue to verify job
completion announcements.