Add concurrent.future.ProcessPoolExecutor as a parallel backend
concurrent.futures.ProcessPoolExecutor is a lot like
multiprocessing.Pool, but it is able to detect when its worker
processes exit unexpectedly. When that happens in
multiprocessing.Pool, the program enters a deadlock and cannot
recover, which is catastrophic for THOR. For concurrent.futures, it at
least raises an exception.

There's no effort here to recover from those exceptions, but at least
the program will terminate.

Crashes of this kind seem to happen mostly when calling pyoorb.