Important Bug Fix: Orbits entirely interior to the observer's orbit will have two possible solutions that intersect the orbit when calculating the topocentric distance.
We select the solution closest to the test orbit's position vector. However,
when the relative observing geometry is such that the two solutions
are nearly parallel (nearly perpendicular to the line of sight which is
itself nearly tangential to the object's orbit) then solutions appear to
be more numerically sensitive. For 2020 AV2, theta_x and theta_y
are typically ~1e-5 degrees (~100 milliarcseconds), however in this instance of observing geometry theta_x degrades to ~1e-4 degrees (~1 arcsec) while theta_y remains the same.